improve delete controller code

+++ backend/src/controllers/Project/slice/file/delete.controller.js
@@ -1,6 +1,7 @@
 import { Model } from "../../../../models/index.js"
 import { deleteFromMinio } from "../../../../utils/deleteFileFromMinio.js";
 import { makeQueue } from "../../../../utils/makeQueue.js";
+import { connection } from '../../../../utils/connection.js'
 
 
 export const deleteFile = async (req, res) => {
@@ -24,9 +25,23 @@ export const deleteFile = async (req, res) => {
             })
         }
 
+        if (file.status === 'deleted') {
+            return res.status(400).json({
+                message: "file Already Deleated"
+            })
+        }
+
         if (file.underProcessing === true) {
-            const filedeleteQueue = new makeQueue('temp-video-delete', { connection })
-            const tempCleanupAt = process.env.NODE_ENV === "production" ? Date.now() + 60 * 60 * 1000 : Date.now() + 6 * 60 * 1000
+
+            await Model.File.updateOne({
+                filename
+            },
+                {
+                    status: "deleted"
+                }
+            )
+            const filedeleteQueue = makeQueue('temp-video-delete', { connection })
+            const tempCleanupAt = process.env.NODE_ENV === "production" ? Date.now() + 3 * 60 * 60 * 1000 : Date.now() + 10 * 60 * 1000
 
             await filedeleteQueue.add(
                 "temp-video-delete",
@@ -43,37 +58,38 @@ export const deleteFile = async (req, res) => {
             );
         }
 
-        const size = file.size;
+        if (file.underProcessing === false) {
 
-        await Model.User.updateOne(
-            {
-                _id: user._id
-            },
-            {
-                $inc: {
-                    storageUsed: -size,
-                    requestsUsed: 1,
+            const size = file.size;
+
+            await Model.User.updateOne(
+                {
+                    _id: user._id
                 },
-            }
-        );
+                {
+                    $inc: {
+                        storageUsed: -size,
+                        requestsUsed: 1,
+                    },
+                }
+            );
 
-        await Model.Project.updateOne(
-            {
-                userid: user._id,
-                _id: project._id,
-            },
-            {
-                $inc: {
-                    requestsUsed: 1,
-                    storageUsed: -size,
+            await Model.Project.updateOne(
+                {
+                    userid: user._id,
+                    _id: project._id,
                 },
-            }
-        );
+                {
+                    $inc: {
+                        requestsUsed: 1,
+                        storageUsed: -size,
+                    },
+                }
+            );
 
-        await deleteFromMinio(`${project}/${filename}`, process.env.MAIN_BUCKET);
-
-        await Model.File.deleteOne({ filename });
-
+            await deleteFromMinio(`${project}/${filename}`, process.env.MAIN_BUCKET);
+            await Model.File.deleteOne({ filename });
+        }
 
         return res.status(200).json({
             message: "Delete Success"
